requirements.lock.txt
beautifulsoup4==4.10.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.1
idna==3.3
importlib-metadata==4.11.3
isort==5.10.1
Markdown==3.3.6
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
requests==2.27.1
soupsieve==2.3.1
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
